BeastLife_AI_Project/requirements.txt
langchain==0.1.14
langgraph==0.0.41
langchain-google-genai==0.0.12
groq==0.4.2
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.5.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiosqlite==3.0.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
